requirements.txt
streamlit==1.31.1
torch==2.1.0
pretrainedmodels==0.7.4
numpy==1.24.3
pillow==10.0.0
streamlit-drawable-canvas==0.9.3
gdown==4.7.1 